Allow disabling URL previews for the webhook
Some IRC channels frequently generate URLs, and the previews on the
Discord side take up large amounts of space that breaks up regular chat.
Bridges not running via a webhook can use Discord channel permissions to
do this, but the webhook messages aren't subject to these permissions.

This commit adds a webhookCanPreviewURLs option that can be set to false
to escape all URLs in messages with the '<>' markdown characters that
suppress previews. It adds a dependency for url-regex 5.0.0 and its
dependencies, ip-regex 4.1.0 and tlds 1.203.0.